researcher_agent/requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
Authlib==1.4.0
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.17.1
azure-mgmt-core==1.6.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-resource==24.0.0
azure-mgmt-subscription==3.0.0
babel==2.18.0
bcrypt==4.3.0
Beaker==1.12.1
beautifulsoup4==4.14.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
blosc2==3.8.0
boto3==1.42.78
botocore==1.42.78
Bottleneck==1.6.0
Brlapi==0.8.7
Brotli==1.2.0
cachetools==5.5.2
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.2
cockpit @ file:///builddir/build/BUILD/cockpit-358-build/cockpit-358/tmp/wheel/cockpit-358-py3-none-any.whl#sha256=f71feeca36dc7d5a2e40b1b62dd015b14ab911c738e799fa2eade9c537009e90
comm==0.2.3
contourpy==1.3.3
crypt_r==3.13.1
cryptography==46.0.6
cssselect==1.3.0
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dataclasses-json==0.6.7
dbus-python==1.4.0
dbus_next==0.2.3
ddgs==9.13.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnf==4.24.0
dnspython==2.8.0
duckduckgo_search==8.1.1
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fedora-third-party==0.10
file-magic==0.4.0
filetype==1.2.0
Flask==3.1.2
flask-babel==4.0.0
Flask-Compress==1.18
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.1.0
gcsfs==2025.9.0
google-api-core==2.27.0
google-api-python-client==2.179.0
google-auth==2.49.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.3.3
google-cloud-storage==2.14.0
google-crc32c==1.8.0
google-genai==1.70.0
google-resumable-media==2.8.0
googleapis-common-protos==1.63.0
greenlet==3.1.1
grpcio==1.48.4
grpcio-status==1.48.4
gssapi==1.7.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.3
humanize==4.13.0
idna==3.10
importlib_resources==6.5.2
inkex==1.4.0
ipykernel==6.29.3
ipython==8.37.0
iso639==0.1.4
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jdcal==1.4.1
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.1
jsonformatter==0.3.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.1
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.7.0
kiwisolver==1.5.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.26
langtable==0.0.69
ldap3==2.9
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
libgravatar==1.0.4
louis==3.33.0
lxml==6.0.1
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.5.0
msal==1.35.1
msal-extensions==1.3.1
msgpack==1.1.0
msrest==0.7.1
multidict==6.6.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.10.0
nest_asyncio==1.6.0
netifaces==0.11.0
nftables==0.1
notebook_shim==0.2.4
numexpr==2.12.1
numpy==2.3.5
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
orjson==3.11.8
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.5
passlib==1.7.4
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
platformdirs==4.2.2
ply==3.11
pooch==1.9.0
primp==1.2.2
productmd==1.50
prometheus_client==0.23.0
prompt_toolkit==3.0.41
propcache==0.4.1
proto-plus==1.22.3
protobuf==3.19.6
psutil==7.0.0
psycopg==3.2.13
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycairo==1.28.0
pycares==4.10.0
pycparser==2.22
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyenchant==3.2.2
Pygments==2.19.1
PyGObject==3.54.5
pyinotify==0.9.6
PyJWT==2.8.0
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
PyMySQL==1.1.2
PyNaCl==1.5.0
pyOpenSSL==26.0.0
pyotp==2.9.0
pyparsing==3.1.2
pyparted==3.13.0
pypng==0.0.21
PyQt5==5.15.12.dev2507081429
PyQt5_sip==12.17.1
pyserial==3.5
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-json-logger==3.3.0
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-socketio==5.16.1
pytz==2026.1
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
pyzmq==27.1.0
qrcode==8.0
QtPy==2.4.3
RapidFuzz==3.12.2
referencing==0.36.2
regex==2026.2.28
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rich==14.1.0
rpds-py==0.27.0
rpm==6.0.1
rsa==4.9
s3transfer==0.16.0
scipy==1.16.2
scour==0.38.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
Send2Trash==1.8.3
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
shellingham==1.5.4
shtab==1.7.2
simple-websocket==1.1.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.11.0
soupsieve==2.8
SQLAlchemy==2.0.48
sqlparse==0.5.3
sshtunnel==0.4.0
stack_data==0.6.3
systemd-python==235
tables==3.10.2
tabulate==0.9.0
tenacity==9.1.4
terminado==0.18.1
tinycss2==1.5.1
tornado==6.5.2
traitlets==5.14.3
typer==0.21.2
typer-cli==0.21.2
typer-slim==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.6.3
user-agents==2.2.0
uuid_utils==0.14.1
wcwidth==0.6.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.0.1
xarray==2025.12.0
xkbregistry==0.3
xlrd==2.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
